+++ NEWS.md
@@ -1,4 +1,5 @@
 # plotme (development version)
+* First version of count_to_sunburst function
 
 # plotme 0.1.0